requirements.txt
autopep8==2.0.2
click==8.1.3
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
pycodestyle==2.10.0
pypng==0.20220715.0
qrcode==7.4.2
SQLAlchemy==2.0.7
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
